Remove dependency on werkzeug
I'm using a patched version of pecan until a new release
is cut with the fix to the test app thing that I submitted
(https://review.openstack.org/#/c/81328/).

There's a couple of changes outside of tests.api - app.py
changes are to fit in better with pecan (specifically my patch),
and the __init__.py change is because I noticed when testing
that there was no `commands` link in the response from /v1.