Bump IPA container to use Trusty
Bump IPA container to use trusty. This gives us a newer version of pip
which is required to install newer requirements that are packaged as
wheels.

Only other change to dockerfile is to ensure python 3 is
removed, in order to avoid raising the size of the container. After this
patch, the agent image goes up in size 3M, which is managable.

Change-Id: I11f39fadb87d42dd15885e6afc63df2c927ce771