Error doc builds on WARN or ERROR
This will ensure that any new commits to IPA won't pass a docs job if
they introduce improperly formatted doc strings. This will enable us to
ensure our generated documentation remains of a high quality.

I will be following this up with a change to project-config to have IPA
run the docs job on every change.

Change-Id: I90236225623298a7070db1ff22acb44dddb1c520