Run IPA in chroot instead of container in CoreOS
systemd-nspawn has been adding more and more security features, the
latest being /sys and /proc/sys being completely read only. This breaks
several things IPA needs to do from the container, including the sysrq
triggers used for reboot.

Additionally, any downstream hardware manager that wants to implement
software RAID requires this fix, as mdadm needs to write to /sys to
assemble arrays. Linux-IO has a similar issue with not operating in
a read only /sys environment.

This changes IPA to run in a chroot instead of a container. This should
be the same from the perspective of IPA, but have no security rules
limiting what we can do inside the ramdisk.

Co-Authored-By: Alex Weeks <alex.weeks@gmail.com>
Change-Id: I6a572bdd791841add913d33c5b79b59033c3f237
Closes-bug: 1524033