Force container builds to use http.us.debian.org
httpredir has multiple mirrors behind it, meaning on any given build
we're relying on several mirrors to all stay up at the same time. This
happens much less often than we'd like.

By forcing this to http.us.debian.org, we're using a more curated set of
mirrors which are less likely to be flakey.

The long term fix is to use a debian mirror provided by infra for
check/gate builds, but until that exists this should help.

Depends-on: I5169faf25c36cfb494c9c1fafb14fb4f7f82cd9e
Related-bug: #1571829
Change-Id: Ia7ea8d909df9ae86a6dbd68ba94746b171535eb8