Probe for TC mirror during tinyipa build
this patch adds a script with known Tinycore Linux HTTP mirrors
and if TINYCORE_MIRROR_URL env var was not provided explicitly,
sets in to the first URL from the list that responds to HEAD request.

The HEAD request is performed without retries and with timeout of 10
seconds.

The list of mirrors is taken from official TC wiki [0].
Currently not responding mirrors are moved to the bottom of the list,
ant those that currently respond with 404 are moved to the very bottom.
The main official repo is left at the first mirror to try.

[0] http://wiki.tinycorelinux.net/wiki:mirrors

Change-Id: I539db06baf1d27b428b75d412851367bfe8a2608