[TinyIPA] Avoid running rm with no parameters
Currently, during ramdisk optimization steps, the find tool is used to
look for files to remove. We pipe all files found by find to the rm
command running as sudo. This is necessary because of the ownership of
the files to be removed.

This commit uses the "--no-run-if-empty" GNU extension to xargs. This
extension should be present on all machines running the TinyIPA build,
because non-GNU systems simply cannot build TinyIPA.

An alternative would have been to use the -delete flag to find, and to
run find as root.

Change-Id: I78854c8901433b7d45e014dd18e7a5ff794a9c45
Closes-bug: #1661219
(cherry picked from commit e49f05028ce9082e46fc0d257ffd0cd53b4a5a78)