Report /dev/disk/by-path on inspection
When inspecting block devices on a node, discover and
report the /dev/disk/by-path/XXX name along with the
/dev/XXX block device name.

The second name does not change between Linux system
reboots and has greater chances to be the same across
similarly configured nodes.

Note: this patch depends on
https://review.openstack.org/#/c/500524/
library patch, but this dependency can't be expressed
with Depends-On clause. Therefore once this patch
requires a followup patch to enable one currently disabled
test in this patch.

Change-Id: I09874f19890500d352521f89573e2aaf50a29022
Closes-Bug: #1679726