Add bindep.txt file
This is used by  [0] to assess if needed system packages are
installed.
It should allow easier installation of Python packages on a clean
system from source and help packagers too.
For now it is just a copy/merge of
files with added platform specifiers.

Packages have been tagged with certain bindep profiles:
- test - needed to compile Python dependencies and for other tox targets,
  profile of this name is used by various gate jobs
- imagebuild - only needed to build various types of deploy ramdisk

Note that DevStack install procedure currently does not use 'bindep',
and still relies on files in  folder.

[0] http://docs.openstack.org/infra/bindep/index.html

Change-Id: I49c9e8314db4f7d778e3316d7f3b9de0fcf89cd3