Fix for broken zuul v3 job and releasenotes
The Zuul v3 job is looking for the log files from the release notes
in: .tox/releasenotes/log/ but we had then in .tox/venv/log/

Fix it to where the Zuul v3 job expects it to be.

Also updated the 'docs' job, even though job isn't failing. If the job
did fail, with current setup, no log files would be copied which would
make debugging more difficult. This change will make it so the log
files are copied.

Change-Id: I5c036db1a9e94ca28e26fd7b45475d52223f5f0e