Use the 'ironic' queue for the gate
When doing the 'gate' build have the ironic-python-agent jobs be part
of the 'ironic' queue. More information on this is in [1]. The benefit
of this is that we are only impacted by ironic projects if we do this.

From [1]:
Pipelines which use the dependent pipeline manager maintain separate
queues for groups of projects. When Zuul serializes a set of changes
which represent future potential project states, it must know about
all of the projects within Zuul which may have an effect on the
outcome of the jobs it runs. If project A uses project B as a library,
then Zuul must be told about that relationship so that it knows to
serialize changes to A and B together, so that it does not merge a
change to B while it is testing a change to A.

[1] https://docs.openstack.org/infra/zuul/feature/zuulv3/user/config.html

Change-Id: Ib0476fb36196d77c431043ea6a48dc842f8d3a65