flake8: Add W503 to ignore list as invalid
W503 according to the [1] pycodestyle docs is not supposed to be
enabled. But it is.

W503 is something we will likely never enable as it is a personal
style decision and can change depending on the code. There is no one
right answer. Interestingly there is also a W504 which is the opposite
check.

[1] http://pycodestyle.pycqa.org/en/latest/intro.html#error-codes

Change-Id: I1025f21a57837e97280f82baba50fdd823a190cc
(cherry picked from commit c09634ad80988e960af2bcaca737be3e87ba7592)