Fix Dockerfile for Debian Jessie
As we learned in I488445dfd261b7bca322a0be7b4d8ca6105750a3
that the the docker file needed to change, this is a targeted
change to address the docker file contents and not update
the coreos version that is built.

This is being done to fix the stable branch builds.

Change-Id: I7251dc7deac788716adad7b596f25fd1633cf289