Change nic driver to e1000 for CoreOS jobs
This change was created after observing a number of the CoreOS
IPA build/test jobs were exhibiting a high failure rate with
the virt-nic.

Similar to the issues recently addressed in ironic's grenade
job, lets explicitly set the NIC to e1000 as that seems to be
more reliable for the CoreOS jobs.

Change-Id: I6365fe138d1b1ddc98a49f1a1da0b82d77678599