Mark CoreOS check jobs non-voting
Due to a higher failure rate, which is largely due
to nested virtualization failures with CoreOS where
one or more of the CI infrastucture providers do not
support nested virtualization, we are marking the
CoreOS CI check jobs as non-voting. This is in order
to reduce the need to recheck.

Change-Id: Id2a5dc766b79ec1131e33f9a49ceca8095c363ef
Story: #2003985
Task: #26928