Add missing dependencies to CoreOS IPA docker image
The iptables and lshw utilities are both required for IPA: iptables is
used when collecting system logs and lshw is used to get the total
physical memory and system details. However, they were not installed in
the docker image embedded in CoreOS IPA images. Since the docker image
is used as a chroot to run IPA, it cannot access binaries available
outside in the main CoreOS file system.

Change-Id: Ic2188b49e717c62315db6c33b74b2e161436b052
Story: #2004092
Task: #27494