Fix download upper constraints
We use the url from tox.ini to download the upper constraints and
recently that has changed to a redirect and curl doesn't
correctly follows redirects by default but stops at the first
page.
Adding -L option to curl command makes curl follows the redirect
until the final location is found and should allow the correct
file to be downloaded.

Change-Id: I25c724ffe189d414643c910be0ef61c550f802b7