Software RAID: Detect component & holder disks in broken RAIDs
This patch amends the component and holder disk detection code
to be able to also extract information from inactive/broken md
devices. This may be required in case of broken RAID-0 (which
will show up as an empty inactive device after a reboot which
none the less keeps its spare devices 'in-use' and prevents
RAID re-creation).

Story: #2006215
Task: #35808
Change-Id: I4ce41b335843e18c67668192d2fccff1c6cdfb92