Fix IPA execution issues due to lack of entropy
While running TinyIPA stable/stein images in KVM virtual machines with
Tenks, we noticed that IPA would start but not execute further, which
appeared to be caused by a lack of entropy.

This patch installs haveged, a simple entropy daemon [1], into the
tinyipa ramdisk and starts it on boot, which resolves the issue.

We first applied this change in ironic-python-agent-builder for master
and stable/train with I7d47b7828733cc03c1ed6c370c451465a6d492b6 [2]. We
still build stable/stein images from the ironic-python-agent repository,
so we can't directly backport the commit. Additionally, we install
haveged from source, because a Tiny Core extension is available only for
x86, not for x86_64 [3]. We use version 1.9.4 as newer versions require
a more recent GNU Automake.

[1] http://www.issihosts.com/haveged/
[2] https://review.opendev.org/#/c/695080/
[3] http://tinycorelinux.net/8.x/x86_64/tcz/

Change-Id: I7d47b7828733cc03c1ed6c370c451465a6d492b6
Story: 2006911
Task: 37554