Stop the API so rescue can exit and succeed
The change in frameworks with the API changed the overall execution
behavior from single threaded serialized execution threaded to
multithreaded through eventlet.

But we signaled ourselves to wait for the API to stop, but never
actually stopped the API.

As setting the flag to stop the API is a hard notification that
the API must now exit, there is no reason to wait and block exit,
hence the simple nature of this change.

Now the threads should exit, IPA should exit, and rescue should
ideally work.

Change-Id: I3faf2d169d4edac6a2321f8649bb9ce478885ea1
Story: 2006998
Task: 37768