Convert jobs to dib
This patch converts the main jobs to dib images.
The discrepancy in the names is due to the fact that some other
projects, like diskimage-builder and ironic-python-agent-builder,
use those jobs, so we leave them for the time being.
The same thing happen for the tinyipa jobs, we create some ad-hoc
jobs until we convert the jobs in the other projects.

Depends-On: https://review.opendev.org/718607

Change-Id: If12a6152015ad00145002f0e5cc2bb948ea6457a