Pin Ironic Tempest Plugin
Pin the ironic-tempest-plugin to an older version so
we don't run the net set of tests with wholedisk + partition
testing.

Also removes CoreOS jobs as CoreOS is EOL and no longer receiving
any updates.

And also signals to the user to try and manually install python-pip
as opposed to install the package from the package manager as
older mirrors have older versions at this point and we are
risking breaking things by automatically doing so. This should
fix test jobs on ubuntu-xenial as there is a conflict in the
repositories.

    python-pip : Depends: python-pip-whl (= 8.1.1-2ubuntu0.4) but 9.0.1-3~3 is to be installed

Change-Id: I6b689bd16dfe6fa849e876f056f1c5cfd1937764